Chore: implement auth / admin refactor
